refactor: utils 생성 후 `set_seed()` 이동
