+++ train.py
@@ -1,4 +1,3 @@
-import random
 from ast import literal_eval
 
 import dotenv
@@ -11,17 +10,7 @@ from peft import LoraConfig
 from transformers import AutoModelForCausalLM, AutoTokenizer
 from trl import DataCollatorForCompletionOnlyLM, SFTConfig, SFTTrainer
 
-
-# 난수 고정
-def set_seed(random_seed):
-    torch.manual_seed(random_seed)
-    torch.cuda.manual_seed(random_seed)
-    torch.cuda.manual_seed_all(random_seed)  # if use multi-GPU
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    np.random.seed(random_seed)
-    random.seed(random_seed)
-
+from utils import set_seed
 
 dotenv.load_dotenv()
 set_seed(42)
